+++ NEWS.md
@@ -2,6 +2,7 @@ X.X (Unreleased)
 ================
 
 * Don't match any files for ROMs or disks with no hashes.
+* When only keeping complete games, minimize searching when a ROM is missing.
 
 2.1 (2024-02-23)
 ================